Use NSDecimalNumber for money calculation

+++ Supermarket/Supermarket/Checkout.h
@@ -18,6 +18,6 @@
 
 - (void)removeAllItems;
 
-- (NSUInteger)totalCost;
+- (double)totalCost;
 
 @end
